requirements.txt
easyocr
pdf2image
pillow
